Add auto connect to 360 eye
